refactor: renamed testing class for InputParser
